[refactor] Move Label/Milestone Buttons Tag to common component
